requirements.txt
anyascii==0.3.2
asgiref==3.7.2
async-timeout==4.0.3
certifi==2023.11.17
charset-normalizer==3.3.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0
django-appconf==1.0.6
django-autocomplete-light==3.9.7
django-bootstrap-input-group==1.0.1
django-bootstrap5==23.3
django-cors-headers==4.4.0
django-filter==23.5
django-guardian==2.4.0
django-import-export==3.3.7
django-modelcluster==6.1
django-permissionedforms==0.1
django-polymorphic==3.1.0
django-redis==5.4.0
django-select2==8.1.2
django-simple-history==3.4.0
django-taggit==4.0.0
django-treebeard==4.7
djangorestframework==3.14.0
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
filetype==1.2.0
hiredis==3.0.0
html5lib==1.1
idna==3.6
l18n==2021.3
Markdown==3.7
MarkupPy==1.14
mysqlclient==2.2.0
odfpy==1.4.1
openpyxl==3.1.2
Pillow==10.1.0
pillow-heif==0.14.0
Pygments==2.18.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tablib==3.5.0
telepath==0.3.1
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
webencodings==0.5.1
xlrd==2.0.1
xlwt==1.3.0
